+++ graphdatascience/model/pipeline_model.py
@@ -1,6 +1,6 @@
 from abc import ABC
 from dataclasses import dataclass
-from typing import Any, Dict
+from typing import Any, Dict, List
 
 from pandas.core.series import Series
 
@@ -50,14 +50,24 @@ class MetricScores(Dict[str, Any]):
         return f"(test={self.test}, outer_train={self.outer_train}, train={self.train}, validation={self.validation})"
 
 
+@dataclass(frozen=True, repr=True)
+class NodePropertyStep:
+
+    proc: str
+    config: Dict[str, Any]
+
+    def __str__(self) -> str:
+        return f"({self.proc}, {self.config})"
+
+
 class PipelineModel(Model, ABC):
     def best_parameters(self) -> "Series[Any]":
         best_params: Dict[str, Any] = self._list_info()["modelInfo"][0]["bestParameters"]
         return Series(best_params)
 
-    def pipeline(self) -> "Series[Any]":
-        steps: Dict[str, Any] = self._list_info()["modelInfo"][0]["pipeline"]
-        return Series(steps)
+    def node_property_steps(self) -> List[NodePropertyStep]:
+        steps: List[Dict[str, Any]] = self._list_info()["modelInfo"][0]["pipeline"]["nodePropertySteps"]
+        return [NodePropertyStep(s["name"], s["config"]) for s in steps]
 
     def metrics(self) -> "Series[Any]":
         model_metrics: Dict[str, Any] = self._list_info()["modelInfo"][0]["metrics"]