+++ graphdatascience/pipeline/nr_training_pipeline.py
@@ -9,7 +9,7 @@ from graphdatascience.pipeline.regression_training_pipeline import (
 )
 
 
-class NRTrainingPipeline(RegressionTrainingPipeline):
+class NRTrainingPipeline(RegressionTrainingPipeline[NRModel]):
     def selectFeatures(self, node_properties: Union[str, List[str]]) -> "Series[Any]":
         query = f"{self._query_prefix()}selectFeatures($pipeline_name, $node_properties)"
         params = {"pipeline_name": self.name(), "node_properties": node_properties}