+++ graphdatascience/tests/integration/test_pipeline_object.py
@@ -6,14 +6,13 @@ from graphdatascience.graph_data_science import GraphDataScience
 from graphdatascience.pipeline.lp_training_pipeline import LPTrainingPipeline
 from graphdatascience.pipeline.nc_training_pipeline import NCTrainingPipeline
 from graphdatascience.pipeline.nr_training_pipeline import NRTrainingPipeline
-from graphdatascience.pipeline.training_pipeline import TrainingPipeline
 from graphdatascience.query_runner.neo4j_query_runner import Neo4jQueryRunner
 
 PIPE_NAME = "pipe"
 
 
 @pytest.fixture
-def pipe(runner: Neo4jQueryRunner, gds: GraphDataScience) -> Generator[TrainingPipeline, None, None]:
+def pipe(runner: Neo4jQueryRunner, gds: GraphDataScience) -> Generator[LPTrainingPipeline, None, None]:
     pipe, _ = gds.beta.pipeline.linkPrediction.create(PIPE_NAME)
 
     yield pipe
@@ -23,19 +22,19 @@ def pipe(runner: Neo4jQueryRunner, gds: GraphDataScience) -> Generator[TrainingP
     runner.run_query(query, params)
 
 
-def test_pipeline_name(pipe: TrainingPipeline) -> None:
+def test_pipeline_name(pipe: LPTrainingPipeline) -> None:
     assert pipe.name() == PIPE_NAME
 
 
-def test_pipeline_type(pipe: TrainingPipeline) -> None:
+def test_pipeline_type(pipe: LPTrainingPipeline) -> None:
     assert pipe.type() == "Link prediction training pipeline"
 
 
-def test_pipeline_creation_time(pipe: TrainingPipeline) -> None:
+def test_pipeline_creation_time(pipe: LPTrainingPipeline) -> None:
     assert pipe.creation_time().year > 2000
 
 
-def test_pipeline_exists(pipe: TrainingPipeline) -> None:
+def test_pipeline_exists(pipe: LPTrainingPipeline) -> None:
     assert pipe.exists()
 
 
@@ -53,11 +52,11 @@ def test_pipeline_drop(gds: GraphDataScience) -> None:
         pipe.drop(failIfMissing=True)
 
 
-def test_pipeline_str(pipe: TrainingPipeline) -> None:
+def test_pipeline_str(pipe: LPTrainingPipeline) -> None:
     assert str(pipe) == "LPTrainingPipeline(name=pipe, type=Link prediction training pipeline)"
 
 
-def test_pipeline_repr(pipe: TrainingPipeline) -> None:
+def test_pipeline_repr(pipe: LPTrainingPipeline) -> None:
     assert "'featureSteps'" in repr(pipe)