+++ graphdatascience/tests/unit/test_pipeline_ops.py
@@ -3,18 +3,17 @@ import pytest
 from .conftest import CollectingQueryRunner
 from graphdatascience.graph_data_science import GraphDataScience
 from graphdatascience.pipeline.lp_training_pipeline import LPTrainingPipeline
-from graphdatascience.pipeline.training_pipeline import TrainingPipeline
 from graphdatascience.server_version.server_version import ServerVersion
 
 PIPELINE_NAME = "dummy"
 
 
 @pytest.fixture
-def pipeline(runner: CollectingQueryRunner, server_version: ServerVersion) -> TrainingPipeline:
+def pipeline(runner: CollectingQueryRunner, server_version: ServerVersion) -> LPTrainingPipeline:
     return LPTrainingPipeline(PIPELINE_NAME, runner, server_version)
 
 
-def test_list_pipelines(runner: CollectingQueryRunner, gds: GraphDataScience, pipeline: TrainingPipeline) -> None:
+def test_list_pipelines(runner: CollectingQueryRunner, gds: GraphDataScience, pipeline: LPTrainingPipeline) -> None:
     gds.beta.pipeline.list(pipeline)
 
     assert runner.last_query() == "CALL gds.beta.pipeline.list($pipeline_name)"
@@ -33,7 +32,7 @@ def test_exists_pipeline(runner: CollectingQueryRunner, gds: GraphDataScience) -
     assert runner.last_params() == {"pipeline_name": "my_pipeline"}
 
 
-def test_drop_pipeline(runner: CollectingQueryRunner, gds: GraphDataScience, pipeline: TrainingPipeline) -> None:
+def test_drop_pipeline(runner: CollectingQueryRunner, gds: GraphDataScience, pipeline: LPTrainingPipeline) -> None:
     gds.alpha.pipeline.drop(pipeline)
 
     assert runner.last_query() == "CALL gds.alpha.pipeline.drop($pipeline_name)"